move API out of main handler
