+++ .changeset/big-ways-hunt.md
@@ -1,0 +1,5 @@
+---
+"@solidjs/start": minor
+---
+
+move API out of main handler
