+++ packages/start/config/server-handler.ts
@@ -94,13 +94,13 @@ async function handleServerFunction(h3Event) {
     ) {
       // workaround for https://github.com/unjs/nitro/issues/1721
       // (issue only in edge runtimes)
-      parsed.push(await new Request(request, { ...request, body: event.node.req.body }).formData());
+      parsed.push(await new Request(request, { ...request, body: h3Event.node.req.body }).formData());
       // what should work when #1721 is fixed 
       // parsed.push(await request.formData);
     } else {
       // workaround for https://github.com/unjs/nitro/issues/1721
       // (issue only in edge runtimes)
-      const tmpReq = new Request(request, { ...request, body: event.node.req.body })
+      const tmpReq = new Request(request, { ...request, body: h3Event.node.req.body })
       // what should work when #1721 is fixed
       // just use request.json() here
       parsed = fromJSON(await tmpReq.json(), {
@@ -138,8 +138,8 @@ async function handleServerFunction(h3Event) {
       }
       // forward headers
       if (result.headers) mergeResponseHeaders(h3Event, result.headers);
-      if (result.customBody) {
-        result = await result.customBody();
+      if ((result as any).customBody) {
+        result = await (result as any).customBody();
       } else if (result.body == undefined) result = undefined;
     }
 
@@ -179,8 +179,8 @@ async function handleServerFunction(h3Event) {
       }
       // forward headers
       if (x.headers) mergeResponseHeaders(h3Event, x.headers);
-      if (x.customBody) {
-        x = await x.customBody();
+      if ((x as any).customBody) {
+        x = await (x as any).customBody();
       } else if (x.body == undefined) x = undefined;
     }
     return x;
