+++ packages/start/server/index.tsx
@@ -1,5 +1,4 @@
 export { StartServer } from "./StartServer";
 export { createHandler } from "./handler";
-export * from "./middleware";
 export * from "./types";
