+++ packages/start/server/spa/index.tsx
@@ -1,4 +1,3 @@
-export * from "../middleware";
 export * from "../types";
 export { StartServer } from "./StartServer";
 export { createHandler } from "./handler";
